manzanillo-qc/requirements.txt
pennylane>=0.36
pennylane-lightning>=0.36
dimod>=0.12
dwave-samplers>=1.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
requests>=2.28
pydantic>=2.0
pyyaml>=6.0
ortools>=9.5
